remove accesslogs.go
