+++ parser_test.go
@@ -7,7 +7,7 @@ import (
 func TestFormatRegex(t *testing.T) {
 	line := `xx.xx.xx.xx - - [24/Jul/2024:00:00:28 +0000] "GET /feed HTTP/1.1" 301 169 "-" "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/109.0.0.0 Safari/537.36"`
 
-	pattern := FormatToRegex(DEFAULT_LOG_FORMAT)
+	pattern := formatToRegex(DEFAULT_LOG_FORMAT)
 	result, err := parseLogLine(pattern, line)
 	assertEqual(t, err, nil)
